fastapi-app/requirements.txt
fastapi
uvicorn
kafka-python
pymongo
